==> mutual_funds_deviations/00_download_index.ipynb <==
import edgar # https://github.com/edgarminers/python-edgar
import pandas as pd
import os

# Save row_data_index
download_directory_root = os.path.join(os.getcwd(), 'data')
download_directory_index = os.path.join(download_directory_root, 'row_data_index')

# Download index of all SEC reports by quarter
since_year=2005
edgar.download_index(download_directory_index, since_year, user_agent='Etienne BRUNO etienne.bruno@epfl.ch', skip_all_present_except_last=False)

def read_tsv(file_name):
    col = ['CIK', 'CNAME', 'FORM_TYPE', 'FILLING_DATE', 'TXT_URL', 'HTML_URL']
    return pd.read_csv(file_name,  sep="|", names=col)

filepaths = sorted(['data/row_data_index/' + path for path in os.listdir(download_directory_index)], reverse=True)
df = pd.concat(map(read_tsv, filepaths), axis=0)
form_to_keep = ['N-CSR', 'N-CSRS']
df = df[df['FORM_TYPE'].isin(form_to_keep)]

# Save to CSV the final dataframe on which we will work
df.to_csv(download_directory_root+"/mutual_funds_data.csv", index=False)

df = pd.read_csv(download_directory_root+"/mutual_funds_data.csv")

df.CIK.unique().size
