==> mutual_funds_deviations/4_benchmark_data.ipynb <==
from IPython.display import display, HTML
display(HTML("<style>.container { width:80% !important; }</style>"))

## GET AVERAGE RETURNS OVER TIME OF THE AVAILABLE FUNDS

# Connection to WRDS
import wrds
db = wrds.Connection(wrds_username='etiennebruno')

# Import CIKs from the dataframe of all cleaned filings
from utils import *
ciks = ', '.join(list(df_from_filings().cik.unique()))

# Get all cik number available (intersection between our dataframe and the data on CRSP)
ciks_data = db.raw_sql(f'''
        select distinct comp_cik
        from crsp_q_mutualfunds.crsp_cik_map
        where comp_cik in ({ciks})
''')
cik_available = list(ciks_data.comp_cik)
cik_available = ', '.join([str(i) for i in cik_available])

# With the mapping table of CRSP, get all companies' fund number
crsp_fundno_available = db.raw_sql(f'''
        select *
        from crsp_q_mutualfunds.crsp_cik_map
        where comp_cik in ({cik_available})
''')
crsp_fundno_available

# Convert the list of fund numbers to a list of string type for future SQL queries
list_float_fundnos = list(crsp_fundno_available.crsp_fundno.unique())
crsp_fundno_available_list_str = ', '.join([str(i) for i in list_float_fundnos])

# Get all available returns for each fund
monthly_returns_all_funds = db.raw_sql(f'''
        select *
        from crsp_q_mutualfunds.monthly_returns
        where crsp_fundno in ({crsp_fundno_available_list_str})
        order by caldt desc
''')
monthly_returns_all_funds

## GET AVERAGE RETURNS OVER TIME OF THE AVAILABLE FUNDS - GROUPED BY MONTH

# Create a new dataframe with some statistic and the average return
monthly_returns_all_funds_aggregated = monthly_returns_all_funds[['caldt', 'mret']].groupby('caldt').agg(
    count=('mret', 'count'),
    mret=('mret', 'mean'),
    mret_max=('mret', 'max'),
    mret_min=('mret', 'min'),
    mret_sdt=('mret', 'std'),
)

monthly_returns_all_funds_aggregated.reset_index(inplace=True)
monthly_returns_all_funds_aggregated['caldt'] = pd.to_datetime(monthly_returns_all_funds_aggregated['caldt'], format= '%Y-%m-%d')

monthly_returns_all_funds_aggregated = monthly_returns_all_funds_aggregated.loc[monthly_returns_all_funds_aggregated['caldt'] >= '2000-01-01']
monthly_returns_all_funds_aggregated

## GET AVERAGE RETURNS OVER TIME OF TRADITIONNAL BENCHMARKS

index_return_query = db.raw_sql("""
    select  
    date,
    vwretd 
    from crsp.msi
    where 
    date>='2000-01-01'""", date_cols=['date']
    )

left = monthly_returns_all_funds_aggregated[['mret', 'caldt']]
right = index_return_query[['date', 'vwretd']]
cumulative_data = pd.merge(left=left, right=right, how='inner', left_on='caldt', right_on='date')

# Compute cumulative returns for both avg of funds and market
cumulative_data['cum_ret_funds'] = (1 + cumulative_data['mret']).cumprod() - 1
cumulative_data['cum_ret_mkt'] = (1 + cumulative_data['vwretd']).cumprod() - 1

## GRAPH OF AVERAGE RETURNS OVER TIME WITH BENCHMARKS

# Rename df for better modularity
data = monthly_returns_all_funds_aggregated.copy()

import matplotlib.pyplot as plt
import numpy as np
import seaborn as sns

%matplotlib inline
%config InlineBackend.figure_format='retina'

sns.set(style="darkgrid")
sns.set_style("whitegrid", {'axes.grid' : False})
sns.set(rc={"figure.figsize":(18, 8)})

plt.tick_params(left=True,
                bottom=True,
                labelleft=True,
                labelbottom=True)

# Plot the average return of all mutual funds
X=data['caldt'].dt.strftime('%Y-%m-%d')
Y=data['mret']
sns.lineplot(x=X, y=Y, linestyle='--')

# Average Value Weighted Index - market
X=index_return_query['date'].dt.strftime('%Y-%m-%d')
Y=index_return_query['vwretd']
sns.lineplot(x=X, y=Y, linestyle='--')

lower_bound = data['mret'] + data['mret_sdt']
upper_bound = data['mret'] - data['mret_sdt']

plt.fill_between(X, lower_bound, upper_bound, alpha=.3)

plt.xlabel('Time')
plt.ylabel('Annual Return')
plt.xticks(X[::12], rotation=45)

plt.legend(['average monthly return of mutual funds', 'value weighted index of the market'])
#plt.savefig('annual_return_plots.png', dpi=200)
#plt.savefig('annual_return_plots_with_market.png', dpi=200)
plt.show()

### GET AND PLOT CUMULATIVE RETURNS

# Rename df for better modularity
data = cumulative_data.copy()

# Plot the cumulative return of all mutual funds
X=data['caldt'].dt.strftime('%Y-%m-%d')
Y=data['cum_ret_funds']*100
sns.lineplot(x=X, y=Y, linestyle='--')

# Cumulative Value Weighted index return - market
X=data['date'].dt.strftime('%Y-%m-%d')
Y=data['cum_ret_mkt']*100
sns.lineplot(x=X, y=Y, linestyle='--')

plt.xlabel('Time')
plt.ylabel('Cumulative return (in %)')
plt.title('Cumulative Return of the average return of the mutual funds and the value weighted market index')
plt.xticks(X[::12], rotation=45)

plt.legend(['Average Performance of mutual funds', 'Value Weighted Market Index'])
#plt.savefig('comparisons_with_vw_market.png', dpi=200)

plt.show()

## FIND VARIABLE NAME IN SCHEMA

%%time
for table in db.list_tables(library='crsp'):
    try:
        l= list(filter(lambda x: ('crsp_fundno' in x) and ('cik' in x), db.get_table(library='crsp_q_mutualfunds', table=table, obs=1).keys()))
    except:
        pass
    if len(l) > 0:
        print(table, "---", l)

%%time
for table in db.list_tables(library='crsp'):
    try:
        tab_keys = db.get_table(library='crsp_q_mutualfunds', table=table, obs=1).keys()
        l = list(filter(lambda x: 'cik' in x, tab_keys))
    except:
        pass
    if len(l) > 0:
        print(table, "---", l, tab_keys)

%%time
for table in db.list_tables(library='crsp'):
    try:
        l= list(filter(lambda x: 'cik' in x, db.get_table(library='crsp', table=table, obs=1).keys()))
    except:
        pass
    if len(l) > 0:
        print(table, "---", l)

## TRY SOME QUERIES

# Connection to WRDS
import wrds
db = wrds.Connection(wrds_username='etiennebruno')

db.list_libraries().sort()
db.list_libraries()

# List the tables of a given library
db.list_tables(library='crsp')

# Get data
daily = db.get_table(library='crsp', table='fund_summary2', obs=10)

# Get funds name over a given period
df_funds_map = db.raw_sql('''
        select distinct on (fund_name) fund_name, ticker, ncusip, cusip8
        from crsp.portnomap
        where begdt <= '2020-01-01' --and enddt >= '2022-01-01'
        order by fund_name asc
''')

# Get funds name over containing some strings of the S&P 500
db.raw_sql('''
        select distinct fund_name
        from crsp.portnomap
        where lower(fund_name::text) LIKE '%%s&p%%'
            and lower(fund_name::text) LIKE '%%500%%'
            --and lower(fund_name::text) LIKE '%%bond%%'
            --and lower(fund_name::text) LIKE '%%national%%'
            --and caldt >= '2010-01-31'
     limit 100
''')

db.raw_sql('''
        select *
        from crsp.fund_summary2
        where fund_name LIKE '%%FundVantage Trust: Gotham Enhanced S&P 500 Index Fund; Institutional Class Shares%%'
        --and caldt >= '2010-01-31'
''')

db.raw_sql('''
        select *
        from crsp.portnomap
        where (fund_name::text) LIKE '%%FundVantage Trust: Gotham Enhanced S&P 500 Index Fund; Institutional Class Shares%%'
        --and caldt >= '2010-01-31'
''')

db.raw_sql('''
        select crsp_fundno, count(mret), max(mret), avg(mret)
        from crsp_q_mutualfunds.monthly_tna_ret_nav
        group by crsp_fundno
        order by count(mret) desc
     limit 10
''')
