==> mutual_funds_deviations/10_fin_bert.ipynb <==
# let us start with importing the dependencies for the analysis
import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import requests
import re
import nltk
from transformers import pipeline

## IMPORT DATAFRAME WITH ALL TEXTS

from utils import *
df = df_from_filings()

df

## IMPORT BERT

# Bert import for sentiment analysis
classifier = pipeline('sentiment-analysis')

# Simple test
result = classifier('We outperformed our benchmark.')
result

## IMPORT FIN-BERT MODEL

# Load Fin-Bert transformer model
fin_classifier = pipeline('sentiment-analysis', model='ProsusAI/finbert')

# Simple test as above with different and more realistic output
result = fin_classifier('We outperformed our benchmark.')
result

# Try with sample of a report filling
txt = df['text'][0]
result = fin_classifier(txt[510:1000])
result

### STOP WORDS LIST COMPLETION

# Load stopwords and save them into a pickle file
with open('stopwords.txt', 'r') as f:
    stop_words = f.read().split('\n')
print(f'Length of initial stop_words : {len(stop_words)}')

# Load nltk stop_words list and add difference to our list
nltk.download('stopwords')
from nltk.corpus import stopwords
cachedStopWords = stopwords.words('english')

for elem in cachedStopWords:
    if elem not in stop_words:
        stop_words.append(elem)
stop_words = sorted(stop_words)
print(f'Length stop_words to be cached as pkl file : {len(stop_words)}')

with open('stop_words_long.pkl', 'wb') as f:
    pickle.dump(stop_words, f)
    
unpickleFile = open('stop_words_long.pkl', 'rb')
stop_words = pickle.load(unpickleFile)
print(f'Length of loaded stop_words list : {len(stop_words)}')

### REMOVE STOP WORDS FROM OUR TEXTS

txt_prep = preprocess(txt)

i=0

i+=1
result = fin_classifier(txt_prep[512*i:512*(i+1)])
result
