==> mutual_funds_deviations/10_regression.ipynb <==
%load_ext ipython_clipboard
%matplotlib inline
import matplotlib.pyplot as plt
import numpy as np
import statsmodels.api as sm
import pandas as pd

df = pd.read_csv('df_final.csv', index_col=0)
df = df[~df.isnull().any(axis=1)]


df

from utils import *
market_returns = load_pkl('market_returns.pkl')
funds_returns = pd.read_csv('funds_returns.csv', index_col=0)

Basic stats

df['outperform_market'] = df.basis_points_delta>0
df['more_positive_than_negative'] = df.finbert_positive > df.finbert_negative

df[['outperform_market', 'more_positive_than_negative']].value_counts()

df.yret.corr(df.vwretd), df.yret.corr(df.finbert_positive), df.yret.corr(df.finbert_negative), df.yret.corr(df.finbert_neutral)

df['finbert_delta'] = df.finbert_positive - df.finbert_negative

First relation (first regression)

%clip
X = df[['vwretd', 'yret']]
# (X - X.mean())/X.std()
# X = sm.add_constant(X)
y = df.finbert_delta

model = sm.OLS(y, X)
results = model.fit()
# print(results.summary())
results.summary().as_latex()

Second relation (second regression)

df1 = df[['comp_cik', 'year', 'vwretd', 'yret']].copy()
df_finbert_predictions = pd.read_csv('df_finbert_predictions.csv', index_col=0)[['cik', 'year', 'finbert_positive', 'finbert_negative', 'finbert_neutral']]
df_finbert_predictions.year -=1
df_finbert_predictions.rename(columns={'cik':'comp_cik'}, inplace=True)
df1 = pd.merge(df1, df_finbert_predictions, on=['year', 'comp_cik'])

%clip
X = df[[ 'vwretd' ,'finbert_positive', 'finbert_negative']]
# X = sm.add_constant(X)
y = df.yret
model = sm.OLS(y, X)
results = model.fit()
print(results.summary())
results.summary().as_latex()

Some tested models that were not put in the report

X = df[['vwretd', 'finbert_positive', 'finbert_negative']]
(X - X.mean())/X.std()
# X = sm.add_constant(X)
y = df.yret

model = sm.OLS(y, X)
results = model.fit()
print(results.summary())

%clip
results.summary().as_latex()

X = df[['vwretd', 'finbert_positive']]
# X = sm.add_constant(X)
y = df.yret
model = sm.OLS(y, X)
results = model.fit()
print(results.summary())

X = df[['vwretd']]
# X = sm.add_constant(X)
y = df.yret
model = sm.OLS(y, X)
results = model.fit()
print(results.summary())

model = sm.OLS(y, X)
results = model.fit()
print(results.summary())

print(sm.OLS(y.shift(1)[1:], X[1:]).fit().summary())

print(sm.OLS(y[1:], X.shift(1)[1:]).fit().summary())

print(sm.OLS(y, df.vwretd).fit().summary())

print(sm.OLS(y.shift(1)[1:], df.vwretd[1:]).fit().summary())
print(sm.OLS(y[1:], df.vwretd.shift(1)[1:]).fit().summary())


X = df[[ 'vwretd' ,'finbert_neutral']]
# X = sm.add_constant(X)
y = df.yret
model = sm.OLS(y, X)
results = model.fit()
print(results.summary())

X = df[[ 'vwretd' ,'finbert_positive']]
X = sm.add_constant(X)
y = df.yret
model = sm.OLS(y, X)
results = model.fit()
print(results.summary())

print(results.summary2())
