==> mutual_funds_deviations/99_convert_pkl_to_csv.ipynb <==
## UTILITIES

### Convert Pickle to CSV

import os
import pandas as pd
import pickle
from utils import save_pkl, load_pkl

folder = 'sentiment_analysis'
path = os.path.join('data', folder)
path_csv = os.path.join('data', folder, 'data_csv')
DS_STORE = '.DS_Store'

files = os.listdir(path)
if DS_STORE in files:
    files.remove(DS_STORE)

# Only keep .pkl files
pickle_files_names = [f for f in files if f.endswith('.pkl')]

# Load all .pkl files and save it under a new folder in .csv
for f in pickle_files_names:

    path_pkl = os.path.join(path, f)
    data = load_pkl(path_pkl)

    if not os.path.exists(path_csv):
        os.makedirs(path_csv)
    
    path_to_save = os.path.join(path_csv, f+'.csv')
    data.to_csv(path_to_save)
